Make dataURItoBlob part of the public API.
